Added function for medium-quality image for small screens and support for background-images

+++ src/progressively.js
@@ -1,10 +1,10 @@
 /*!
- * progressively 1.1.2
- * https://github.com/thinker3197/progressively
- * @license MIT licensed
- *
- * Copyright (C) 2016-17 Ashish
- */
+* progressively 1.1.3
+* https://github.com/thinker3197/progressively
+* @license MIT licensed
+*
+* Copyright (C) 2016-17 Ashish
+*/
 
 ;
 (function (root, factory) {
@@ -22,7 +22,7 @@
 
   var progressively = {}
 
-  var defaults, poll, onLoad, inodes
+  var defaults, poll, onLoad, inodes, sminodes
 
   onLoad = function () {}
 
@@ -32,12 +32,22 @@
       o[prop] = secondaryObject.hasOwnProperty(prop) ? secondaryObject[prop] : primaryObject[prop]
     }
     return o
-  };
+  }
 
+/**
+ * Checks, if element is hidden
+ * @param  object DOMElement
+ * @return {Boolean}    [description]
+ */
   function isHidden (el) {
     return (el.offsetParent === null)
-  };
+  }
 
+/**
+ * Check if element is currently visible
+ * @param  object DOMElement
+ * @return boolean
+ */
   function inView (el) {
     if (isHidden(el)) {
       return false
@@ -52,7 +62,7 @@
     do {
       box = el.getBoundingClientRect()
 
-      if (top <= box.bottom === false) {
+      if (!(top <= box.bottom)) {
         return false
       }
       if ((top + height) <= box.top) {
@@ -63,24 +73,54 @@
     } while (el !== document.body)
 
     return top <= document.documentElement.clientHeight
-  };
+  }
 
-  function loadImage (el) {
+/**
+ * Load image and add loaded-class. Loads the minified version, if small display
+ * @param  object DOMElement
+ * @param  object defaults
+ * @return boolean true, if fully loaded; false, if minified version was loaded
+ */
+  function loadImage (el, defaults) {
     setTimeout(function () {
       var img = new Image()
 
       img.onload = function () {
         el.classList.remove('progressive--not-loaded')
         el.classList.add('progressive--is-loaded')
-        el.src = this.src
+
+        if (el.classList.contains('progressive__bg')) {
+// Load image as css-background-image
+          el.style['background-image'] = 'url("' + this.src + '")'
+        } else {
+          el.src = this.src
+        }
 
         onLoad(el)
       }
 
+// Load minified version, if viewport-width is smaller than defaults.smBreakpoint:
+      if (getClientWidth() < defaults.smBreakpoint && el.getAttribute('data-progressive-sm')) {
+        el.classList.add('progressive--loaded-sm')
+        img.src = el.getAttribute('data-progressive-sm')
+      }
+
+      el.classList.remove('progressive--loaded-sm')
       img.src = el.getAttribute('data-progressive')
     }, defaults.delay)
-  };
+  }
 
+/**
+ * Returns the width of the client's viewport
+ * @return integer client-width
+ */
+  function getClientWidth () {
+    return Math.max(document.documentElement.clientWidth, window.innerWidth || 0)
+  }
+
+/**
+ * Listens to an event, and throttles
+ */
   function listen () {
     if (poll) {
       return
@@ -92,17 +132,22 @@
       poll = null
     }, defaults.throttle)
   }
-    /*
-     * default settings
-     */
 
+ /*
+ * default settings
+ */
   defaults = {
     throttle: 300, // appropriate value, don't change unless intended
     delay: 100,
     onLoadComplete: function () {},
-    onLoad: function () {}
+    onLoad: function () {},
+    smBreakpoint: 600
   }
 
+/**
+ * Initializer. Finds image-elements and adds listeners.
+ * @param  object options
+ */
   progressively.init = function (options) {
     options = options || {}
 
@@ -110,46 +155,74 @@
 
     onLoad = defaults.onLoad || onLoad
 
-    inodes = [].slice.call(document.querySelectorAll('.progressive__img'))
+    inodes = [].slice.call(document.querySelectorAll('.progressive__img, .progressive__bg'))
+    sminodes = []
 
     progressively.render()
 
     if (document.addEventListener) {
       root.addEventListener('scroll', listen, false)
+      root.addEventListener('resize', listen, false)
       root.addEventListener('load', listen, false)
     } else {
       root.attachEvent('onscroll', listen)
+      root.attachEvent('onresize', listen)
       root.attachEvent('onload', listen)
     }
   }
 
+/**
+ * Loads necessary images in small or full quality.
+ */
   progressively.render = function () {
     var elem
 
     for (var i = inodes.length - 1; i >= 0; --i) {
       elem = inodes[i]
 
-      if (inView(elem) && elem.classList.contains('progressive--not-loaded')) {
-        loadImage(elem)
+      if (inView(elem) && (elem.classList.contains('progressive--not-loaded') || elem.classList.contains('progressive--loaded-sm'))) {
+        loadImage(elem, defaults)
+        if (elem.classList.contains('progressive--loaded-sm')) {
+          sminodes.push(elem)
+        }
         inodes.splice(i, 1)
+      }
+    }
+
+    if (getClientWidth() >= defaults.smBreakpoint) {
+      for (var j = sminodes.length - 1; j >= 0; --j) {
+        elem = sminodes[j]
+
+        if (inView(elem) && (elem.classList.contains('progressive--not-loaded') || elem.classList.contains('progressive--loaded-sm'))) {
+          loadImage(elem, defaults)
+          sminodes.splice(i, 1)
+        }
       }
     }
 
     this.check()
   }
 
+/**
+ * Check if all images are loaded in full quality, then drop.
+ */
   progressively.check = function () {
-    if (!inodes.length) {
+    if (!inodes.length && !sminodes.length) {
       defaults.onLoadComplete()
       this.drop()
     }
   }
 
+/**
+ * Drops progressively-listeners
+ */
   progressively.drop = function () {
     if (document.removeEventListener) {
       root.removeEventListener('scroll', listen)
+      root.removeEventListener('resize', listen)
     } else {
       root.detachEvent('onscroll', listen)
+      root.detachEvent('onresize', listen)
     }
     clearTimeout(poll)
   }
